+++ cvpr/ml/rf/src/SplitNode.cpp
@@ -14,22 +14,22 @@ void SplitNodeShapeIndexed<ty>::init_params(const TrainingSet &train_set, const 
 	assert(nullptr != param_);
 	
 	std::uniform_int_distribution<std::size_t>	distr_idx(0, param_->num_shape);
-	std::uniform_int_distribution<int>			distr_pos(-param_->radius, param_->radius);
+	std::uniform_int_distribution<double>		distr_pos(-param_->radius, param_->radius);
 
 	// 対象形状点決める
 	shape_index	=	distr_idx(rnd);
 
 	// 相対特徴点位置を決める
 	for (int ii = 0; ii < NUM_FEATURE_POS; ++ii) {
-		offsets.push_back(cv::Point2d(dist_pos(rnd), distr_pos(rnd)));
+		offsets.push_back(cv::Point2f(dist_pos(rnd), distr_pos(rnd)));
 	}
 }
 
 template<typename ty>
 double SplitNodeShapeIndexed<ty>::kernel_function(const cv::Mat &feature, const SplitNodeParameterBase *param) const
 {
-	std::vector<cv::Point2d>	warped_offsets;
-	std::vector<cv::Point2d>	feature_pos;
+	std::vector<cv::Point2f>	warped_offsets;
+	std::vector<cv::Point2f>	feature_pos;
 	cv::Mat	transform;
 	const ShapeIndexedSplitParameter	*param_	=	dynamic_cast<const ShapeIndexedSplitParameter*>(param);
 	std::vector<double>	val;
@@ -39,7 +39,7 @@ double SplitNodeShapeIndexed<ty>::kernel_function(const cv::Mat &feature, const 
 	cv::perspectiveTransform(offsets, warped_offsets, param_->transform);
 
 	for (std::size_t ii = 0; ii < warped_offsets.size(); ++ii) {
-		cv::Point2d	pos	=	param_->shape[shape_index] + warped_offsets[ii];
+		cv::Point2f	pos	=	param_->shape[shape_index] + warped_offsets[ii];
 
 		pos	=	round(feature, pos); // 範囲外なら例外でいい気もする?
